requirements.txt
django
djangorestframework
adrf
django-cors-headers
python-dotenv
pymemcache
gunicorn
uvicorn
schedule
httpx
dico-api
faker
githubkit[auth-app]
django-oauth-toolkit
django-allauth
